Move logger to outer
